requirements-test.txt
factory-boy==2.9.2
flake8==3.4.1
isort==4.2.15
mypy==0.521
pytest-cov==2.5.1
pytest-django==3.1.2
